arquitecturaPractica/lakehouse/requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dynaconf==3.2.5
exceptiongroup==1.2.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
pur==7.3.1
py4j==0.10.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.2
rsa==4.9
six==1.16.0
tomli==2.0.1
tzdata==2024.1
urllib3==2.2.1